remove dupe initializeExecutor
